moved all prop types to props.d.ts, renamed TodoList.interface.ts to types.d.ts
