+++ src/components/ThemePalette/ThemePalette.tsx
@@ -4,7 +4,7 @@ import { Palette } from 'react-bootstrap-icons';
 import ColorDotList from './ColorDotList';
 import ThemeButtons from './ThemeButtons';
 
-import { ThemePaletteProps } from '../TodoList.interface';
+import { ThemePaletteProps } from '../props';
 
 export default function ThemePalette(props: ThemePaletteProps) {
     const { paletteOpen, themeColors, themeMode, togglePaletteOpen, handleChangeColor, setMode } = props;
