+++ src/hooks/hooks.d.ts
@@ -1,4 +1,4 @@
-import { Task } from '../components/TodoList.interface';
+import { Task } from '../components/types';
 import { KEYS } from './constants';
 
 export type Key = typeof KEYS[keyof typeof KEYS];
